% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_fns.R
\name{gamma_cos}
\alias{gamma_cos}
\title{Create a historical functional coefficient where risk is of the form \eqn{c \times cos(2 \times \pi \times u/p)} with \eqn{u=s-t} for a fixed landmark time}
\usage{
gamma_cos(t, s, period = 1/2, scale = 10)
}
\arguments{
\item{t}{vector containing the historical time for evaluating the historical functional coefficient}

\item{s}{scalar containing the "landmark" time for evaluating the historical functional coefficient}

\item{period}{period of the cosine function (\eqn{p})}

\item{scale}{scaling factor for the coefficient (\eqn{c})}
}
\description{
Create a historical functional coefficient where risk is of the form \eqn{c \times cos(2 \times \pi \times u/p)} with \eqn{u=s-t} for a fixed landmark time
}
