% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_fns.R
\name{gamma_exp_decay}
\alias{gamma_exp_decay}
\title{Create a historical functional coefficient using an exponential decay model for a fixed landmark time}
\usage{
gamma_exp_decay(t, s, d = 1/5, scale = 1)
}
\arguments{
\item{t}{vector containing the historical time for evaluating the historical functional coefficient}

\item{s}{scalar containing the "landmark" time for evaluating the historical functional coefficient}

\item{d}{numeric indicating the mean of the exponential random variable whose density function determins the rate of decay for the coefficient function}

\item{scale}{scaling factor for the coefficient}
}
\description{
Create a historical functional coefficient using an exponential decay model for a fixed landmark time
}
