% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_fns.R
\name{gen_gamma_mat}
\alias{gen_gamma_mat}
\title{Function for evaluating the historical functional coefficient. Called by the genSurv function}
\usage{
gen_gamma_mat(gammafun, S, Tind, ...)
}
\arguments{
\item{gammafun}{function which evaluates the landmark historical functional coefficient. Should evaluate for a fixed landmark time and a vector of historical times using arguments passed down tot he function via ...}

\item{S}{vector containing the "landmark" times for evaluating the historical functional coefficient}

\item{Tind}{vector containing the historical time for evaluating the historical functional coefficient}

\item{...}{arguments to be passed to gammafun}

\item{period}{period of the cosine function (\eqn{p})}

\item{scale}{scaling factor for the coefficient (\eqn{c})}
}
\description{
Function for evaluating the historical functional coefficient. Called by the genSurv function
}
