% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_fns.R
\name{genSurv}
\alias{genSurv}
\title{Function for simulating survival curves for the historical functional Cox model using a single time-fixed and time-varying covariate}
\usage{
genSurv(
  lambda0,
  beta,
  gamma = gamma_exp_decay,
  tind,
  Z,
  X,
  N,
  cens_fun = NULL,
  returnS = FALSE,
  ...
)
}
\arguments{
\item{lambda0}{function which takes time as an input and returns the baseline hazard}

\item{beta}{function which takes time as an input and returns a potentially time varying effect for a single scalar covariate}

\item{gamma}{function which takes time and historical time as input and returns a historical functional coefficient}

\item{tind}{grid on which to evaluate the survival curve}

\item{Z}{matrix of the longitudinal/time varying predictor to be used in creating the survival curve for individuals, evaluated on the grid supplied to the "tind" argument (should have N rows)}

\item{X}{vector of time-fixed covariate (should be of length N)}

\item{N}{integer, number of functions observed. Should be equal to the length of X and number of rows of Z}

\item{cens_fun}{function which generates censoring times}

\item{returnS}{logical indicating whether to return the simulated survival functions along with survival times}

\item{...}{arguments to be passed to gen_gamma_mat}
}
\description{
Function for simulating survival curves for the historical functional Cox model using a single time-fixed and time-varying covariate
}
