% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_fns.R
\name{getSurvT}
\alias{getSurvT}
\title{Simulate survival times using a survival curve evaluated on a discrete grid}
\usage{
getSurvT(survfunction, tind)
}
\arguments{
\item{survfunction}{vector containing estimates of the survival function \eqn{S(t) \in [0,1]}}

\item{tind}{vector indicating the time points associated with survfunction input}
}
\description{
Simulate survival times using a survival curve evaluated on a discrete grid
}
