% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria_fns.R
\name{ID_AUC}
\alias{ID_AUC}
\title{Function for calculating "local" AUC(s,t)}
\usage{
ID_AUC(marker, Stime, status, predict.time, entry = NULL)
}
\arguments{
\item{marker}{vector of predicted survival probability at the time indicated by the predict.time argument}

\item{Stime}{vector of observed survival times for participants}

\item{status}{vector of event indicators for participants}

\item{predict.time}{scalar indicating the time (t) AUC(s,t) should be calcuated at}

\item{entry}{exclude participants with events/censoring before this time}
}
\description{
Function for calculating "local" AUC(s,t)
}
