% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_fns.R
\name{make_lm_data}
\alias{make_lm_data}
\title{Function for creating the landmark dataset}
\usage{
make_lm_data(
  data,
  vars_tv,
  vars_tf,
  id,
  event_time,
  time,
  S,
  wide = FALSE,
  censor_var,
  w = NULL
)
}
\arguments{
\item{data}{frame of data (in long format)}

\item{vars_tv}{character vector of time-varying predictors}

\item{vars_tf}{character vector of time-fixed predictors (must not vary within grouping variable)}

\item{id}{character vector of length 1 indicating grouping ("id") variable}

\item{event_time}{character vector of length 1 indicating event time variable ("event_time")}

\item{time}{character vector of length 1 indicating time of observation of the time-varying predictors ("time"). Currently only supports a common grid of observation times if there are multiple time-varying predictors.}

\item{S}{numeric vector of landmark times (S)}

\item{wide}{logical indicating whether data supplied to the function are in wide format (FALSE) or long format (TRUE)}

\item{censor_var}{character vector of length 1 indicating event indicator variable. This variable must be coded as 1=event, 0=censor}

\item{w}{numeric vector of prediction windows (w). Must be either length 1 or same length as the vector of landmark times (s)}
}
\description{
Function for creating the landmark dataset
}
