% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_fns.R
\name{predict_lm_cumulative_hazard}
\alias{predict_lm_cumulative_hazard}
\title{Function for creating the landmark dataset}
\usage{
predict_lm_cumulative_hazard(fit, s, tpred)
}
\arguments{
\item{fit}{fitter LM-HFLCM}

\item{s}{length 1 numeric vector indicating landmark time}

\item{tpred}{numeric vector of predictor times to evaluate the cumulative hazard function}
}
\description{
Function for creating the landmark dataset
}
